backend/requirements.txt
Django==3.2.13
djangorestframework==3.12.4
psycopg2-binary==2.8.6
python-dotenv==0.20.0
pytz==2022.1
pytest==7.1.2
pytest-django==4.5.2
sqlparse==0.4.2
Pillow==9.2.0
djoser==2.1.0
django_extra_fields==3.0.2
django-filter==21.1
drf_yasg==1.21.3
django-cors-headers==3.13.0
gunicorn==20.0.4
reportlab==3.6.9
